Refactoring
Remove AbstractOpsLoggerFactory, and move its logic into
ConfigurationInfo. Remove AsyncOpsLoggerFactory and
BasicOpsLoggerFactory's superclass, and have them obtain
the significant ops logger dependencies (Destinations,
including StackTraceProcessors, error handlers, correlation
id suppliers) they need to construct opsloggers directly from
the ConfigurationInfo instance they have a reference to.
