+++ opslogger/src/main/java/com/equalexperts/logging/impl/BasicOpsLoggerFactory.java
@@ -10,17 +10,18 @@ import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-public class BasicOpsLoggerFactory extends AbstractOpsLoggerFactory {
+public class BasicOpsLoggerFactory {
+
+    private final ConfigurationInfo configurationInfo;
 
     public BasicOpsLoggerFactory(ConfigurationInfo configurationInfo) {
-        super(configurationInfo);
+        this.configurationInfo = configurationInfo;
     }
 
     public <T extends Enum<T> & LogMessage> OpsLogger<T> build() throws IOException {
-        Supplier<Map<String,String>> correlationIdSupplier = configureCorrelationIdSupplier();
-        Consumer<Throwable> errorHandler = configureErrorHandler();
-        Destination<T> destination = configureDestination();
+        Supplier<Map<String,String>> correlationIdSupplier = configurationInfo.configureCorrelationIdSupplier();
+        Consumer<Throwable> errorHandler = configurationInfo.configureErrorHandler();
+        Destination<T> destination = configurationInfo.configureDestination();
         return new BasicOpsLogger<>(Clock.systemUTC(), correlationIdSupplier, destination, new ReentrantLock(), errorHandler);
     }
-
 }
